db changed (sql server => in memory), query tests are done

+++ WebApi/Application/ActorOperations/Commands/DeleteActor/DeleteActorCommand.cs
@@ -18,6 +18,9 @@ namespace WebApi.Application.ActorOperations.Commands.DeleteActor
             if (actor is null)
                 throw new InvalidOperationException("No actor to delete was found.");
 
+            if (actor.MovieActors != null && actor.MovieActors.Any())
+                throw new InvalidOperationException("Cannot delete this actor because he/she have any movie.");
+
             _context.Actors.Remove(actor);
             _context.SaveChanges();
         }
